Remove duplicating tests
